+++ ci/rebuild-libcabi-realloc.sh
@@ -43,7 +43,7 @@ version=$(grep '^version =' ./Cargo.toml | sed 's/.*"\(.*\)"/\1/' | sed 's/\./_/
 
 sym=cabi_realloc_wit_bindgen_$version
 
-cat >./crates/guest-rust/src/cabi_realloc.rs <<-EOF
+cat >./crates/guest-rust/rt/src/cabi_realloc.rs <<-EOF
 // This file is generated by $0
 
 #[no_mangle]
@@ -53,11 +53,11 @@ pub unsafe extern "C" fn $sym(
     align: usize,
     new_len: usize,
 ) -> *mut u8 {
-    crate::rt::cabi_realloc(old_ptr, old_len, align, new_len)
+    crate::cabi_realloc(old_ptr, old_len, align, new_len)
 }
 EOF
 
-cat >./crates/guest-rust/src/cabi_realloc.c <<-EOF
+cat >./crates/guest-rust/rt/src/cabi_realloc.c <<-EOF
 // This file is generated by $0
 
 #include <stdint.h>
@@ -70,16 +70,16 @@ void *cabi_realloc(void *ptr, size_t old_size, size_t align, size_t new_size) {
 }
 EOF
 
-rm -f crates/guest-rust/src/cabi_realloc.o
-$WASI_SDK_PATH/bin/clang crates/guest-rust/src/cabi_realloc.c \
-  -O -c -o crates/guest-rust/src/cabi_realloc.o
+rm -f crates/guest-rust/rt/src/cabi_realloc.o
+$WASI_SDK_PATH/bin/clang crates/guest-rust/rt/src/cabi_realloc.c \
+  -O -c -o crates/guest-rust/rt/src/cabi_realloc.o
 
 # Remove the `producers` section. This appears to differ whether the host for
 # clang is either macOS or Linux. Not needed here anyway, so discard it to help
 # either host produce the same object.
-wasm-tools strip -d producers ./crates/guest-rust/src/cabi_realloc.o \
-  -o ./crates/guest-rust/src/cabi_realloc.o
+wasm-tools strip -d producers ./crates/guest-rust/rt/src/cabi_realloc.o \
+  -o ./crates/guest-rust/rt/src/cabi_realloc.o
 
-rm -f crates/guest-rust/src/libwit_bindgen_cabi_realloc.a
-$WASI_SDK_PATH/bin/llvm-ar crus crates/guest-rust/src/libwit_bindgen_cabi_realloc.a \
-  crates/guest-rust/src/cabi_realloc.o
+rm -f crates/guest-rust/rt/src/libwit_bindgen_cabi_realloc.a
+$WASI_SDK_PATH/bin/llvm-ar crus crates/guest-rust/rt/src/libwit_bindgen_cabi_realloc.a \
+  crates/guest-rust/rt/src/cabi_realloc.o
