+++ crates/guest-rust/macro/src/lib.rs
