+++ crates/guest-rust/rt/src/cabi_realloc.rs
@@ -7,5 +7,5 @@ pub unsafe extern "C" fn cabi_realloc_wit_bindgen_0_20_0(
     align: usize,
     new_len: usize,
 ) -> *mut u8 {
-    crate::rt::cabi_realloc(old_ptr, old_len, align, new_len)
+    crate::cabi_realloc(old_ptr, old_len, align, new_len)
 }
